requirements.txt
flask==2.3.3
python-dotenv==1.0.0
dashscope>=1.14.0
websockets==11.0.3
Werkzeug==2.3.7 